# backend/api-spec.yaml
openapi: 3.0.3
info:
  title: Test Case Manager API
  description: RESTful API for Test Case Management System with TestLink integration
  version: 1.0.0
  contact:
    name: Test Case Manager
    url: http://192.168.4.121:3000
servers:
  - url: http://192.168.4.121:3001/api
    description: Production server
  - url: http://localhost:3001/api
    description: Development server
paths:
...
  /testsuites:
    get:
      summary: List test suites
      tags: [Test Suites]
      parameters:
        - name: project_id
          in: query
          description: Filter by project ID
          schema:
            type: integer
        - name: parent_suite_id
          in: query
          description: Filter by parent test suite ID
          schema:
            type: integer
      responses:
        '200':
          description: Test suites retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestSuite'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a new test suite
      tags: [Test Suites]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, name]
              properties:
                project_id:
                  type: integer
                name:
                  type: string
                  maxLength: 255
                description:
                  type: string
                parent_suite_id:
                  type: integer
                  nullable: true
      responses:
        '201':
          description: Test suite created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/TestSuite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /testsuites/{id}:
    get:
      summary: Get test suite by ID
      tags: [Test Suites]
      parameters:
        - $ref: '#/components/parameters/TestSuiteId'
      responses:
        '200':
          description: Test suite retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/TestSuite'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update test suite
      tags: [Test Suites]
      parameters:
        - $ref: '#/components/parameters/TestSuiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 255
                description:
                  type: string
                parent_suite_id:
                  type: integer
                  nullable: true
      responses:
        '200':
          description: Test suite updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/TestSuite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete test suite
      tags: [Test Suites]
      parameters:
        - $ref: '#/components/parameters/TestSuiteId'
      responses:
        '200':
          description: Test suite deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
...
    TestSuiteId:
      name: id
      in: path
      required: true
      description: Test Suite ID
      schema:
        type: integer
...
  schemas:
...
    TestSuite:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        name:
          type: string
        description:
          type: string
        parent_suite_id:
          type: integer
          nullable: true
        external_id:
          type: string
          nullable: true
        node_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...